refactor: move Schwab error decorator (#128)
