+++ src/open_stocks_mcp/tools/responses.py
@@ -64,22 +64,6 @@ def handle_robin_stocks_sync_errors(
     return wrapper
 
 
-def handle_schwab_errors(
-    func: Callable[P, Awaitable[R]],
-) -> Callable[P, Awaitable[R | dict[str, Any]]]:
-    """Decorator to handle Schwab API errors consistently."""
-
-    @functools.wraps(func)
-    async def wrapper(*args: P.args, **kwargs: P.kwargs) -> R | dict[str, Any]:
-        context = f"in {func.__name__}"
-        try:
-            return await func(*args, **kwargs)
-        except Exception as e:
-            return create_error_response(e, context)
-
-    return wrapper
-
-
 def sanitize_api_response(data: Any) -> Any:
     """Sanitize API response data to remove sensitive information."""
     if isinstance(data, dict):
