+++ src/open_stocks_mcp/tools/schwab/__init__.py
@@ -1,0 +1,1 @@
+"""Schwab-specific tool helpers."""
