+++ src/open_stocks_mcp/tools/schwab_options_tools.py
@@ -14,8 +14,8 @@ from open_stocks_mcp.tools.broker_utils import (
 from open_stocks_mcp.tools.error_handling import (
     create_error_response,
     create_success_response,
-    handle_schwab_errors,
 )
+from open_stocks_mcp.tools.schwab.error_handling import handle_schwab_errors
 
 _OPEN_OPTION_STATUSES = frozenset(
     {
