+++ tests/unit/test_error_handling.py
@@ -26,12 +26,12 @@ from open_stocks_mcp.tools.error_handling import (
     create_success_response,
     execute_with_retry,
     handle_robin_stocks_errors,
-    handle_schwab_errors,
     log_api_call,
     sanitize_api_response,
     validate_period,
     validate_symbol,
 )
+from open_stocks_mcp.tools.schwab.error_handling import handle_schwab_errors
 
 pytestmark = [pytest.mark.unit, pytest.mark.journey_system]
